refactor(synapsys): reduce complexity in replay CLI to satisfy quality gate
Split synapsys-replay.js into focused sibling modules (replay-events,
replay-aggregate, replay-judge, replay-report) and extract helper
functions to bring each function under the cyclomatic/cognitive limits
and the file under the 400-line cap. Behavior unchanged — all 34
integration tests pass and the quality gate reports no new violations.
