+++ plugins/synapsys/lib/replay-aggregate.js
@@ -1,0 +1,142 @@
+'use strict';
+
+/**
+ * synapsys-replay — per-memory aggregation + tightening heuristic. Extracted
+ * to keep the CLI entrypoint under the 400-line quality cap.
+ *
+ * Public surface (re-exported by scripts/synapsys-replay.js):
+ *   - splitTopLevelAlternation(triggerPrompt)
+ *   - fpRate(relevant, irrelevant)
+ *   - aggregateReport(tuples, judgments)
+ *   - suggestTightening(memory, agg)
+ */
+
+function updateDepth(depth, ch) {
+  if (ch === '(') depth.paren++;
+  else if (ch === ')') depth.paren = Math.max(0, depth.paren - 1);
+  else if (ch === '[') depth.bracket++;
+  else if (ch === ']') depth.bracket = Math.max(0, depth.bracket - 1);
+}
+
+/**
+ * Split `trigger_prompt` on top-level `|` only (outside `(...)` / `[...]`).
+ */
+function splitTopLevelAlternation(triggerPrompt) {
+  if (typeof triggerPrompt !== 'string' || triggerPrompt.length === 0) return [];
+  const arms = [];
+  const depth = { paren: 0, bracket: 0 };
+  let buf = '';
+  for (let i = 0; i < triggerPrompt.length; i++) {
+    const ch = triggerPrompt[i];
+    updateDepth(depth, ch);
+    if (ch === '|' && depth.paren === 0 && depth.bracket === 0) {
+      arms.push(buf);
+      buf = '';
+      continue;
+    }
+    buf += ch;
+  }
+  arms.push(buf);
+  return arms;
+}
+
+/**
+ * fp_rate = 1 - relevant / (relevant + irrelevant). judge_failed excluded.
+ */
+function fpRate(relevant, irrelevant) {
+  const denom = relevant + irrelevant;
+  if (denom <= 0) return null;
+  return 1 - relevant / denom;
+}
+
+function newReportEntry() {
+  return {
+    fires: 0,
+    relevant: 0,
+    irrelevant: 0,
+    judge_failed: 0,
+    fp_rate: null,
+    sample_matches: [],
+    _hasUps: false,
+    _hasPtu: false,
+  };
+}
+
+function recordFire(entry, bucket, t) {
+  entry.fires += 1;
+  bucket.events.add(t.event);
+  if (t.event === 'UserPromptSubmit') entry._hasUps = true;
+  if (t.event === 'PreToolUse') entry._hasPtu = true;
+  if (t.matched_substring) {
+    bucket.subs.set(t.matched_substring, (bucket.subs.get(t.matched_substring) || 0) + 1);
+  }
+}
+
+function applyJudgmentToEntry(entry, j) {
+  if (j && entry._hasUps) {
+    entry.relevant = j.relevant || 0;
+    entry.irrelevant = j.irrelevant || 0;
+    entry.judge_failed = j.judge_failed || 0;
+    entry.fp_rate = fpRate(entry.relevant, entry.irrelevant);
+  } else if (!entry._hasUps && entry._hasPtu) {
+    entry.relevant = null;
+    entry.irrelevant = null;
+    entry.fp_rate = null;
+  }
+}
+
+function topSampleMatches(subs) {
+  return Array.from(subs.entries())
+    .sort((a, b) => b[1] - a[1] || a[0].localeCompare(b[0]))
+    .slice(0, 3)
+    .map(([s]) => s);
+}
+
+/**
+ * Aggregate per-memory metrics from replay tuples + judge results.
+ * PTU-only memories report `relevant=null` / `fp_rate=null` (spec §Decision).
+ * `sample_matches` is the top-3 most-frequent distinct matched substrings.
+ */
+function aggregateReport(tuples, judgments) {
+  const report = {};
+  const fireBuckets = {};
+  for (const t of tuples) {
+    if (!report[t.memory_name]) {
+      report[t.memory_name] = newReportEntry();
+      fireBuckets[t.memory_name] = { events: new Set(), subs: new Map() };
+    }
+    if (t.fired) recordFire(report[t.memory_name], fireBuckets[t.memory_name], t);
+  }
+  for (const name of Object.keys(report)) {
+    const entry = report[name];
+    applyJudgmentToEntry(entry, judgments && judgments[name]);
+    entry.sample_matches = topSampleMatches(fireBuckets[name].subs);
+    delete entry._hasUps;
+    delete entry._hasPtu;
+  }
+  return report;
+}
+
+/**
+ * Heuristic R8 — emit advisory when `fp_rate > 0.70` AND `trigger_prompt`
+ * contains short (<=5 chars) single-word alternation arms.
+ */
+function suggestTightening(memory, agg) {
+  if (!memory || !agg) return null;
+  if (agg.fp_rate === null || agg.fp_rate === undefined) return null;
+  if (!(agg.fp_rate > 0.7)) return null;
+  const arms = splitTopLevelAlternation(memory.triggerPrompt || '');
+  const candidates = arms.filter((a) => {
+    const trimmed = a.trim();
+    return trimmed.length > 0 && trimmed.length <= 5 && /^[A-Za-z0-9_-]+$/.test(trimmed);
+  });
+  if (candidates.length === 0) return null;
+  return { memory: memory.name, candidates };
+}
+
+module.exports = {
+  splitTopLevelAlternation,
+  fpRate,
+  aggregateReport,
+  suggestTightening,
+};
