+++ plugins/synapsys/lib/replay-report.js
@@ -1,0 +1,117 @@
+'use strict';
+
+/**
+ * synapsys-replay — report renderers (extracted from the CLI script to keep
+ * the main entrypoint under the 400-line quality cap).
+ *
+ * Public surface (re-exported by scripts/synapsys-replay.js):
+ *   - renderJson(agg, suggestions, meta)
+ *   - renderReport(agg, suggestions, meta)
+ */
+
+// Cost model constants (R17): ~500 input + 5 output tokens per judge call.
+// Haiku pricing approx $1.00 / 1M input + $5.00 / 1M output tokens.
+const PRICE_PER_TOKEN = 1e-6; // ~$1/1M as a conservative per-token blend.
+const TOKENS_PER_JUDGE_CALL = 500 + 5;
+
+/**
+ * Render the aggregated report as machine-readable JSON (R9, G6).
+ *
+ * Spec §Security — `ANTHROPIC_API_KEY` is never included in output; this
+ * function only serialises the inputs it is handed (no env access).
+ */
+function renderJson(agg, suggestions, meta) {
+  const memories = Object.keys(agg).map((name) => {
+    const m = agg[name];
+    return {
+      name,
+      fires: m.fires,
+      relevant: m.relevant,
+      irrelevant: m.irrelevant,
+      judge_failed: m.judge_failed,
+      fp_rate: m.fp_rate,
+      sample_matches: m.sample_matches,
+    };
+  });
+  const payload = {
+    memories,
+    suggestions: Array.isArray(suggestions) ? suggestions : [],
+    events_total: meta && typeof meta.events_total === 'number' ? meta.events_total : 0,
+    events_ups: meta && typeof meta.events_ups === 'number' ? meta.events_ups : 0,
+    events_ptu: meta && typeof meta.events_ptu === 'number' ? meta.events_ptu : 0,
+  };
+  return JSON.stringify(payload, null, 2);
+}
+
+function renderHeaderLine(m) {
+  return (
+    `store=${m.store || ''} window=${m.window || ''} events=${m.events_total || 0} ` +
+    `UPS=${m.events_ups || 0} PTU=${m.events_ptu || 0}`
+  );
+}
+
+function renderTableHeader() {
+  return (
+    'Memory'.padEnd(30) +
+    'Fires'.padStart(7) +
+    'Relevant'.padStart(10) +
+    'FP%'.padStart(8) +
+    '  Sample matches'
+  );
+}
+
+function formatRelevant(value) {
+  return value === null || value === undefined ? '—' : String(value);
+}
+
+function formatFpPct(value) {
+  return value === null || value === undefined ? '—' : `${Math.round(value * 100)}%`;
+}
+
+function renderMemoryRow(name, e) {
+  const samples = (e.sample_matches || []).slice(0, 3).join(', ');
+  return (
+    name.padEnd(30) +
+    String(e.fires).padStart(7) +
+    formatRelevant(e.relevant).padStart(10) +
+    formatFpPct(e.fp_rate).padStart(8) +
+    '  ' +
+    samples
+  );
+}
+
+function renderSuggestionsBlock(suggestions) {
+  const sugs = Array.isArray(suggestions) ? suggestions : [];
+  if (sugs.length === 0) return ['  (none)'];
+  return sugs.map(
+    (s) => `  - ${s.memory}: tighten short arms [${(s.candidates || []).join(', ')}]`
+  );
+}
+
+function renderCostFooter(judgeCalls) {
+  const cost = TOKENS_PER_JUDGE_CALL * judgeCalls * PRICE_PER_TOKEN;
+  return ['', `est. cost ≈ $${cost.toFixed(4)} (${judgeCalls} judge calls)`];
+}
+
+/**
+ * Render the human-readable report (R11). Includes header, per-memory table
+ * rows, Suggestions section, and cost footer when `meta.judgeCalls > 0` (R17).
+ */
+function renderReport(agg, suggestions, meta) {
+  const m = meta || {};
+  const lines = [renderHeaderLine(m), '', renderTableHeader()];
+  for (const name of Object.keys(agg)) {
+    lines.push(renderMemoryRow(name, agg[name]));
+  }
+  lines.push('', 'Suggestions:');
+  lines.push(...renderSuggestionsBlock(suggestions));
+  if (m.judgeCalls && m.judgeCalls > 0) {
+    lines.push(...renderCostFooter(m.judgeCalls));
+  }
+  return lines.join('\n') + '\n';
+}
+
+module.exports = {
+  renderJson,
+  renderReport,
+};
